+++ History.md
@@ -1,5 +1,6 @@
 # History
 
 - 0.X.X
-    - ***0.2.0*** (2018-09-13) - Highlight major Forex news events on investing.com. Set the time automatically to NYC time.
+    - ***0.2.1*** (2018-09-19) - [investing.com] Filter to 2nd and 3rd importance events. Add a new GBP event.
+    - ***0.2.0*** (2018-09-13) - [investing.com] Highlight major Forex news events. Set the time automatically to NYC time.
     - ***0.1.0*** (2018-09-03) - Initial release